# docs/swagger.yaml
basePath: /
definitions:
  entity.Ids:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  entity.Products:
    properties:
      amount:
        type: integer
      product_id:
        type: integer
      storage_id:
        type: integer
    type: object
info:
  contact: {}
  description: API Server for SHOP
  title: SHOP API
  version: "1.0"
paths:
  /v1/amount/{storage}:
    get:
      consumes:
      - application/json
      description: Take amounts products
      parameters:
      - description: storage id
        in: path
        name: storage
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Products'
            type: array
        "500":
          description: Internal Server Error
      summary: Take amounts products
      tags:
      - Products
  /v1/release:
    post:
      consumes:
      - application/json
      description: Release some products
      parameters:
      - description: products id
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Ids'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Release products
      tags:
      - Products
  /v1/reserve:
    post:
      consumes:
      - application/json
      description: Reserve some products
      parameters:
      - description: products id
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Ids'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Reserve products
      tags:
      - Products
swagger: "2.0"